fix: Split tests
Baseline tests for when random values need mocked, end to end, measuring the performance of the actual recurrent net.
